move callback arg parsing into dependencies.py
